% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbh.est-data.R
\docType{data}
\name{lbh.est}
\alias{lbh.est}
\title{Extended selection table of \emph{Phaethornis longirostris} songs}
\format{One object of class extended selection table (following warbleR package format) and data frame.}
\usage{
data(lbh.est)
}
\description{
'lbh.est' contains an \emph{extented selection table} following the warbleR package object format. The data contains the annotations and wave objects of 50 long-billed hermit songs from 10 different song types (5 per song type, 'lek.song.type' column) recorded at 8 leks ('lek' column) in the Caribbean slope of Costa Rica.
}
\details{
Extended selection table of \emph{Phaethornis longirostris} (Long-billed Hermit) songs.
}
\keyword{datasets}
