% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monk.parakeet.est-data.R
\docType{data}
\name{monk.parakeet.est}
\alias{monk.parakeet.est}
\title{Extended selection table of monk parakeet contact calls}
\format{One object of class extended selection table (following warbleR package format) and data frame.}
\source{
Data used in Smith-Vidaurre, Araya-Salas & Wright (2019)
}
\usage{
data(monk.parakeet.est)
}
\description{
'monk.parakeet.est' contains an \emph{extented selection table} following the warbleR package object format. The data contains the annotations and wave objects of 52 monk parake calls from four different social scales (flock individual, pair, site, 'scale' column). The 'ID' column labels selections belonging to the same individual when `scale == "Individual`. See Smith-Vidaurre, Araya-Salas & Wright (2019) for more details.
}
\details{
Extended selection table of monk parakeet (\emph{Myiopsitta monachus}) contact calls.
}
\references{
Smith-Vidaurre G, Araya-Salas M, & Wright T F. 2019. Individual signatures outweigh social group identity in contact calls of a communally nesting parrot.  Behavioral Ecology.
}
\keyword{datasets}
