% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatureSounds-package.R
\docType{package}
\name{NatureSounds}
\alias{NatureSounds-package}
\alias{NatureSounds}
\title{AnimalSounds: Animal Sounds for Bioacustic Analysis}
\description{
NatureSounds is a package containing a collection of example animal sounds for bioacoustic analysis. 
The main goal is to provide example recordings to faciliate the development of bioacoustic 
tools in the R environment.
}
\details{
Most data sets are provided as extended selection tables. These are R objects that contains both the annotations (locations of signals in time and frequency) and the corresponding acoustic data as wave objects (\href{https://marce10.github.io/warbleR/articles/annotation_data_format.html#extended-selection-tables}{see a full description here}). Therefore, these are **self-contained objects** since the original sound files are no longer needed to perform acoustic analyses. These objects are created by the function `selection_table()` from \href{https://cran.r-project.org/package=warbleR}{warbleR}.

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maRce10/NatureSounds}
  \item Report bugs at \url{https://github.com/maRce10/NatureSounds/issues}
}

}
\author{
Marcelo Araya-Salas
Maintainer: Marcelo Araya-Salas (\email{marceloa27@gmail.com})
}
