% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thyroptera.est-data.R
\docType{data}
\name{thyroptera.est}
\alias{thyroptera.est}
\title{Extended selection table of Spix's disc-winged bat social calls}
\format{
One object of class extended selection table (following warbleR package format) and data frame.
}
\source{
Data used in Araya-Salas et al (2020)
}
\usage{
data(thyroptera.est)
}
\description{
'thyroptera.est' contains an \emph{extented selection table} following the warbleR package object format. The data contains the annotations and wave objects of 32 Spix's disc-winged bats social calls from 8 individuals ('id' column) belonging to 4 social groups ('group' column). See Araya-Salas et al (2020) for more details. Can be used (among other things) to test for individual and group acoustic signatures.
}
\details{
Extended selection table of Spix's disc-winged bats (\emph{Thyroptera}) social calls.
}
\references{
Araya-Salas M, A Hernandez-Pinson, N Rojas, G Chaverri. 2020. Ontogeny of an interactive call-and-response system in Spix's disc-winged bats. Animal Behaviour.
}
\keyword{datasets}
