% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatureSounds-package.R
\docType{package}
\name{NatureSounds}
\alias{NatureSounds}
\alias{NatureSounds-package}
\title{AnimalSounds: Animal Sounds for Bioacustic Analysis}
\description{
NatureSounds is a package containing a collection of example animal sounds for bioacoustic analysis.
The main goal is to provide example recordings to faciliate the development of bioacoustic 
tools in the R environment. When possible, selection tables with the time/frequency coordinates of the sounds 
in the example recordings will also be included. Note that this package does 
not contain any function. Check out the packages 'warbleR', 'tuneR' and 'seewave' 
for sound file manipulation and analysis in R.
}
\details{
License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas
  
  Maintainer: Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
