% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatureSounds-package.R
\docType{package}
\name{NatureSounds}
\alias{NatureSounds}
\alias{NatureSounds-package}
\title{AnimalSounds: Animal sounds for bioacustic analysis}
\description{
NatureSounds is a package containing a collection of example animal sounds for bioacoustic analysis.
The main goal is to provide example recordings of animal sounds to faciliate the development of bioacoustic 
tools in the R environment. When possible selection tables with the time/frequency coordinates of the sounds 
in the example recordings will also be included.
}
\details{
License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas
  
  Maintainer: Marcelo Araya-Salas (\email{araya-salas@cornell.edu})
}
