% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phae.long.est-data.R
\docType{data}
\name{Phae.long.est}
\alias{Phae.long.est}
\title{Extended selection table of long-billed hermit songs}
\format{One object of class extended selection table (following warbleR package format) and data frame.}
\source{
Data used in Araya-Salas et al (2019)
}
\usage{
data(Phae.long.est)
}
\description{
'Phae.long.est' contains an \emph{extented selection table} following the warbleR package object format. The data contains the annotations and wave objects of 100 long-billed hermit songs from 10 different song types (10 per song type, 'lek.song.type' column) recorded at 8 leks ('lek' column) in the Caribbean slope of Costa Rica.
}
\details{
Extended selection table of Long-billed Hermit (\emph{Phaethornis longirostris}) songs.
}
\references{
Araya-Salas M, G Smith-vidaurre,  D Mennill, P Gonzalez-Gomez, J Cahill & T Wright. 2019. Social group signatures in hummingbird displays provide evidence of co-occurrence of vocal and visual learning. Proceedings of the Royal Society B. 286: 20190666
}
\keyword{datasets}
